Organized code into packages finally.
I started off making the plugin like this, but since it was all off the top of my head, I didn't know where to look for bugs. So I had to write it all with factory methods to be sure it would work. Now It can be organized a litte better.